Fix cast warnings for cast_unittest on win_x64
Fixed a compiler warning on win_x64 with size_t assigned to int.

R=justinlin@chromium.org

Review URL: https://codereview.chromium.org/25727008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@226821 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/cast/framer/frame_buffer.h
@@ -22,7 +22,7 @@ class FrameBuffer {
   FrameBuffer();
   ~FrameBuffer();
   void InsertPacket(const uint8* payload_data,
-                    int payload_size,
+                    size_t payload_size,
                     const RtpCastHeader& rtp_header);
   bool Complete() const;
 
@@ -41,7 +41,7 @@ class FrameBuffer {
   uint16 max_packet_id_;
   uint16 num_packets_received_;
   bool is_key_frame_;
-  int total_data_size_;
+  size_t total_data_size_;
   uint8 last_referenced_frame_id_;
   uint32 rtp_timestamp_;
   PacketMap packets_;
